Moved batch_summary.py to folder - types
