+++ trolley/__init__.py
@@ -6,5 +6,5 @@ from trolley.batch import Batch
 from trolley.payment import Payment
 from trolley.recipient_account import RecipientAccount
 from trolley.gateway import Gateway
-from trolley.batch_summary import BatchSummary
+from trolley.types.batch_summary import BatchSummary
 from trolley.types.log import Log
